+++ src/chinodeco/__init__.py
@@ -1,7 +1,7 @@
 # !/usr/bin/env Python3
 # -*- coding:utf-8 -*-
 
-__version__ = "0.0.3"
+__version__ = "0.0.4"
 
 from .base import decochain
 from .parameter import (